+++ gamemodes/terrortown/gamemode/shared/sh_main.lua
@@ -182,6 +182,12 @@ hook.Add("PreRegisterSWEP", "TTT2RegisterSWEP", TTT2RegisterSWEP)
 -- @hook
 -- @realm shared
 function GM:TTT2Initialize()
+	-- Migrate all changes of TTT2 for the new version
+	if migrations.MigrateToVersion(GAMEMODE.Version) then
+		-- OnSuccess Update the database to the new version
+		versions.UpdateDatabase()
+	end
+
 	-- load all roles
 	roles.OnLoaded()